add delete prompt

+++ static/controllers/careerResourceController.js
@@ -33,6 +33,7 @@ qccApp.controller("CareerResourceController", ["$scope", "$location", "$window",
 
 
   $scope.deleteResource = function(resource){
+    if (confirm('Are you sure you want to delete this resource?')) {
       $http.defaults.headers.delete = { "Content-Type": "application/json;charset=utf-8" };//NEED THIS
       console.log(resource);
       $http.delete("/careerResource", {data: resource})//NEED THIS FORMAT TOO
@@ -42,6 +43,9 @@ qccApp.controller("CareerResourceController", ["$scope", "$location", "$window",
       }, function(error){
           console.log(error);
       });
+    } else {
+    }
+
   };
 
   $scope.toggleResourceDesc = function(event) {
